Move PushRequestInterceptor class to PushRequestInterceptorTest
